touch CF directories during startup
We try to be robust against files disappearing (due to any kind of corruption)
inside the data directory.

But if the data directory itself goes missing, that's a situation that we don't
handle correctly.  We will keep accepting writes normally, but when we try to
flush the memtable to disk, we'll fail with a system error.

Having the CF directory disappearing is not a common thing. But it is also one
that we can easily protect against, by touching all CF directories we know
about on startup.

Fixes #999

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <ed66373dccca11742150a6d08e21ece3980227d3.1457379853.git.glauber@scylladb.com>